requirements.txt
ocpp==2.1.0
websockets~=13.1
pytest~=8.3.3
pytest-asyncio~=0.24.0
pyhumps~=3.8.0
jsonschema~=4.23.0
pytest-env